backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
overrides==3.0.0
dataclasses==0.6
python-dotenv==0.13.0
PyDispatcher==2.0.5
requests==2.22.0
PyYAML==5.3.1
